# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "HouseHold Service API",
    "description": "Application to serve the needs for household errands",
    "version": "1.0.0-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8080/household-errands-service",
      "description": "Local server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "tags": [
    {
      "name": "Auth-Controller",
      "description": "Authentication Controller"
    },
    {
      "name": "Household-Controller",
      "description": "Household Controller"
    },
    {
      "name": "Backlog-Controller",
      "description": "Backlog Controller"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Auth-Controller"],
        "summary": "Login with email and password",
        "operationId": "login",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": ["Auth-Controller"],
        "summary": "Register a new user",
        "operationId": "register",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Email already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/household": {
      "get": {
        "tags": ["Household-Controller"],
        "summary": "Get all households",
        "operationId": "getAllHouseholds",
        "responses": {
          "200": {
            "description": "Successfully retrieved all households",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HouseholdResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "tags": ["Household-Controller"],
        "summary": "Create a new household",
        "operationId": "createHousehold",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Household created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/household/{householdId}": {
      "get": {
        "tags": ["Household-Controller"],
        "summary": "Get a household by ID",
        "operationId": "getHouseholdById",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved household",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Household not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Household-Controller"],
        "summary": "Update a household",
        "operationId": "updateHousehold",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HouseholdRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Household updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HouseholdResponse"
                }
              }
            }
          },
          "404": {
            "description": "Household not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Household-Controller"],
        "summary": "Delete a household",
        "operationId": "deleteHousehold",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Household deleted"
          },
          "404": {
            "description": "Household not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/household/{householdId}/members": {
      "get": {
        "tags": ["Household-Controller"],
        "summary": "Get all members of a household",
        "operationId": "getMembers",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved members",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MemberResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Household not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/household/{householdId}/backlog/item/{itemId}": {
      "get": {
        "tags": ["Backlog-Controller"],
        "summary": "Get an item",
        "operationId": "getItem",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found an item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Backlog-Controller"],
        "summary": "Remove an item",
        "operationId": "removeItem",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully removed an item"
          },
          "404": {
            "description": "Item not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/household/{householdId}/backlog/list": {
      "get": {
        "tags": ["Backlog-Controller"],
        "summary": "Get a backlog",
        "operationId": "getBacklog",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found the backlog",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Backlog"
                }
              }
            }
          },
          "404": {
            "description": "Backlog not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/household/{householdId}/backlog/item": {
      "post": {
        "tags": ["Backlog-Controller"],
        "summary": "Add an item",
        "operationId": "addItem",
        "parameters": [
          {
            "name": "householdId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Item"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully added an item",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Item"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "example": "john.smith@example.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJqb2huLnNtaXRoQGV4YW1wbGUuY29tIn0.abc123"
          },
          "email": {
            "type": "string",
            "example": "john.smith@example.com"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Smith"
          },
          "householdId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName"],
        "properties": {
          "email": {
            "type": "string",
            "example": "jane.doe@example.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "householdId": {
            "type": "integer",
            "format": "int64",
            "example": 1,
            "description": "Optional: link user to an existing household at registration"
          }
        }
      },
      "HouseholdRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Smith Household"
          }
        }
      },
      "HouseholdResponse": {
        "type": "object",
        "title": "Household data model",
        "properties": {
          "householdId": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Smith Household"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemberResponse"
            }
          }
        }
      },
      "MemberResponse": {
        "type": "object",
        "title": "Member data model",
        "properties": {
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "The member's user ID",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "The member's email address",
            "example": "john.smith@example.com"
          },
          "firstName": {
            "type": "string",
            "description": "The member's first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "The member's last name",
            "example": "Smith"
          }
        }
      },
      "Item": {
        "type": "object",
        "title": "Item data model",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int64",
            "description": "The Id representing an item",
            "example": 1
          },
          "itemName": {
            "type": "string",
            "description": "The name representing an item",
            "example": "Toaster"
          }
        }
      },
      "Backlog": {
        "type": "object",
        "title": "Backlog data model",
        "properties": {
          "itemsList": {
            "type": "array",
            "description": "A list of items",
            "items": {
              "$ref": "#/components/schemas/Item"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "title": "Error Response",
        "required": ["statusCode", "message"],
        "properties": {
          "statusCode": {
            "type": "string",
            "description": "The status code representing the error",
            "example": "500"
          },
          "message": {
            "type": "string",
            "description": "A message providing context about the causes of the error",
            "example": "An error has occurred"
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "title": "Success Response",
        "required": ["statusCode", "message"],
        "properties": {
          "statusCode": {
            "type": "string",
            "description": "The status code representing the success",
            "example": "200"
          },
          "message": {
            "type": "string",
            "description": "A message providing context about the success of the post request",
            "example": "Successfully added a team member"
          }
        }
      }
    }
  }
}