# src/main/java/co/za/household/domain/model/HouseholdRequest.java
package co.za.household.domain.model;

public class HouseholdRequest {

    private String name;

    public String getName() {
        return name;
    }

    public HouseholdRequest setName(String name) {
        this.name = name;
        return this;
    }
}