# src/main/java/co/za/household/domain/model/LoginRequest.java
package co.za.household.domain.model;

public class LoginRequest {

    private String email;
    private String password;

    public String getEmail() {
        return email;
    }

    public LoginRequest setEmail(String email) {
        this.email = email;
        return this;
    }

    public String getPassword() {
        return password;
    }

    public LoginRequest setPassword(String password) {
        this.password = password;
        return this;
    }
}